Udpate service-import tests
